+++ internal/cmd/render.go
@@ -3,16 +3,16 @@ package cmd
 import (
 	"bufio"
 	"bytes"
-	"fmt"
 	"os"
 	"regexp"
 
 	"github.com/mvisonneau/s5/pkg/cipher"
+	"github.com/pkg/errors"
 	log "github.com/sirupsen/logrus"
 	"github.com/urfave/cli/v2"
 )
 
-// Render is used for the render commands
+// Render is used for the render commands.
 func Render(ctx *cli.Context) (int, error) {
 	cipherEngine, err := getCipherEngine(ctx)
 	if err != nil {
@@ -26,17 +26,20 @@ func Render(ctx *cli.Context) (int, error) {
 	if ctx.NArg() > 1 ||
 		(ctx.String("output") != "" && ctx.Bool("in-place")) {
 		if err = cli.ShowSubcommandHelp(ctx); err != nil {
-			return 1, err
+			return 1, errors.Wrap(err, "rendering subcommand help")
 		}
-		return 1, fmt.Errorf("Invalid arguments")
+
+		return 1, errors.New("invalid arguments")
 	}
 
 	var fi *os.File
+
 	if ctx.NArg() == 1 {
 		log.Debugf("Opening input file : %s", ctx.Args().First())
+
 		fi, err = os.Open(ctx.Args().First())
 		if err != nil {
-			return 1, err
+			return 1, errors.Wrap(err, "opening input file")
 		}
 	} else {
 		log.Debug("Reading from stdin")
@@ -49,55 +52,64 @@ func Render(ctx *cli.Context) (int, error) {
 	var buf bytes.Buffer
 
 	log.Debug("Starting deciphering")
+
 	for in.Scan() {
 		buf.WriteString(
 			re.ReplaceAllStringFunc(in.Text(), func(src string) string {
 				log.Debugf("found: %v", re.FindStringSubmatch(src)[1])
+
 				plain, err := cipherEngine.Decipher(re.FindStringSubmatch(src)[1])
 				if err != nil {
 					panic(err)
 				}
+
 				return plain
 			}) + "\n")
 	}
 
 	if err = fi.Close(); err != nil {
-		return 1, err
+		return 1, errors.Wrap(err, "closing input file")
 	}
 
 	if err = in.Err(); err != nil {
-		return 1, err
+		return 1, errors.Wrap(err, "reading input file")
 	}
 
 	var fo *os.File
-	if ctx.String("output") != "" {
+
+	switch {
+	case ctx.String("output") != "":
 		log.Debugf("Creating and outputing to file : %s", ctx.String("output"))
+
 		fo, err = os.Create(ctx.String("output"))
 		if err != nil {
-			return 1, err
+			return 1, errors.Wrap(err, "creating output file")
 		}
+
 		defer closeFile(fo)
-	} else if ctx.Bool("in-place") {
+	case ctx.Bool("in-place"):
 		log.Debug("Updating the source file (in-place)")
+
 		fo, err = os.Create(ctx.Args().First())
 		if err != nil {
-			return 1, err
+			return 1, errors.Wrap(err, "writing output file")
 		}
+
 		defer closeFile(fo)
-	} else {
+	default:
 		log.Debug("Outputing to stdout")
+
 		fo = os.Stdout
 	}
 
 	out := bufio.NewWriter(fo)
-	_, err = out.Write(buf.Bytes())
-	if err != nil {
-		return 1, err
+
+	if _, err = out.Write(buf.Bytes()); err != nil {
+		return 1, errors.Wrap(err, "writing output file")
 	}
 
-	err = out.Flush()
-	if err != nil {
-		return 1, err
+	if err = out.Flush(); err != nil {
+		return 1, errors.Wrap(err, "flushing output file")
 	}
 
 	return 0, nil