+++ pkg/cipher/gcp/gcp.go
@@ -6,29 +6,30 @@ import (
 	"fmt"
 
 	cloudkms "cloud.google.com/go/kms/apiv1"
-	kmspb "google.golang.org/genproto/googleapis/cloud/kms/v1"
-
+	"github.com/pkg/errors"
 	log "github.com/sirupsen/logrus"
+	kmspb "google.golang.org/genproto/googleapis/cloud/kms/v1"
 )
 
-// Config handles necessary information for AES
+// Config handles necessary information for AES.
 type Config struct {
 	KmsKeyName string
 }
 
-// Client is an handler for encryption functions
+// Client is an handler for encryption functions.
 type Client struct {
 	*cloudkms.KeyManagementClient
 	Context *context.Context
 	Config  *Config
 }
 
-// NewClient configures a client for encryption purposes
+// NewClient configures a client for encryption purposes.
 func NewClient(config *Config) (*Client, error) {
 	ctx := context.Background()
+
 	client, err := cloudkms.NewKeyManagementClient(ctx)
 	if err != nil {
-		return nil, err
+		return nil, errors.Wrap(err, "creating new client")
 	}
 
 	return &Client{
@@ -38,9 +39,10 @@ func NewClient(config *Config) (*Client, error) {
 	}, nil
 }
 
-// Cipher : Cipher a value using the provided key
+// Cipher : Cipher a value using the provided key.
 func (c *Client) Cipher(value string) (string, error) {
 	log.Debug("Ciphering using a GCP KMS key")
+
 	req := &kmspb.EncryptRequest{
 		Name:      c.Config.KmsKeyName,
 		Plaintext: []byte(value),
@@ -48,18 +50,19 @@ func (c *Client) Cipher(value string) (string, error) {
 
 	resp, err := c.Encrypt(*c.Context, req)
 	if err != nil {
-		return "", err
+		return "", errors.Wrap(err, "ciphering value")
 	}
 
 	return base64.StdEncoding.EncodeToString(resp.Ciphertext), nil
 }
 
-// Decipher : Decipher a value using the TransitKey
+// Decipher : Decipher a value using the TransitKey.
 func (c *Client) Decipher(value string) (string, error) {
 	log.Debugf("Deciphering '%s' using a GCP KMS key", value)
+
 	ciphertext, err := base64.StdEncoding.DecodeString(value)
 	if err != nil {
-		return "", fmt.Errorf("base64decode error : %s - value : %s", err, value)
+		return "", errors.Wrap(err, fmt.Sprintf("base64decode error, input : '%s'", value))
 	}
 
 	req := &kmspb.DecryptRequest{
@@ -69,7 +72,7 @@ func (c *Client) Decipher(value string) (string, error) {
 
 	resp, err := c.Decrypt(*c.Context, req)
 	if err != nil {
-		return "", err
+		return "", errors.Wrap(err, "deciphering value")
 	}
 
 	return string(resp.Plaintext), nil