+++ pkg/cipher/vault/vault_test.go
@@ -1,3 +1,3 @@
-package vault
+package vault_test
 
 // TODO: Implement!